requirements.txt
openai>=1.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.20.0
datasets>=2.14.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
transformers>=4.30.0
ipython>=8.0.0 